requirements.txt
langchain
deepagents
python-dotenv
langchain-openai
langchain-text-splitters
beautifulsoup4
pinecone
pyTelegramBotAPI